# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InOrder API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "AuthHeader": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header"
      },
      "RefreshHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "refreshToken"
      }
    }
  },
  "paths": {
...
    "/api/items": {
      "get": {
        "summary": "Get items",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Initial offset for pagination",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create new item",
        "description": "Need to have authorization of CHEF or above to perform this request",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/items/{itemid}": {
      "get": {
        "summary": "Get specific item by itemID",
        "parameters": [
          {
            "in": "path",
            "name": "itemid",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of item to get"
          }
        ],
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "summary": "Delete item",
        "description": "Need to have authorization of CHEF or above to perform this request",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "name": "itemid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "summary": "Update an existing item's properties",
        "description": "Need to have authorization of CHEF or above to perform this request",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "name": "itemid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "required": false
                  },
                  "description": {
                    "type": "string",
                    "required": false
                  },
                  "price": {
                    "type": "number",
                    "required": false
                  },
                  "updtags": {
                    "type": "array",
                    "required": false,
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/items/upload": {
      "post": {
        "summary": "Upload image of an item",
        "description": "Need to have authorization of CHEF or above to perform this request",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/items/tags": {
      "get": {
        "summary": "Get all existing item tags",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{orderid}/items": {
      "get": {
        "summary": "Get all ordered items in a specific order",
        "description": "If you have the role of CUSTOMER then you must have issued the order to see its items",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderid",
            "in": "path",
            "type": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Order a new item",
        "description": "If you have the role of CUSTOMER then you must have issued the order to order an item in it",
        "security": [
          {
            "AuthHeader": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderid",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item_id": {
                    "required": true,
                    "type": "number"
                  },
                  "quantity": {
                    "required": true,
                    "type": "number"
                  },
                  "instructions": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}